Renames Write.* to CmdWrite.* to better reflect the division of labour between those files and WriteQuestion.*.
