+++ quizzer.vcxproj
@@ -106,7 +106,7 @@
     <ClCompile Include="Play.cpp" />
     <ClCompile Include="Question.cpp" />
     <ClCompile Include="util.cpp" />
-    <ClCompile Include="Write.cpp" />
+    <ClCompile Include="CmdWrite.cpp" />
     <ClCompile Include="WriteFlashcard.cpp" />
     <ClCompile Include="WriteQuestion.cpp" />
   </ItemGroup>
@@ -120,7 +120,7 @@
     <ClInclude Include="Play.h" />
     <ClInclude Include="Question.h" />
     <ClInclude Include="util.h" />
-    <ClInclude Include="Write.h" />
+    <ClInclude Include="CmdWrite.h" />
     <ClInclude Include="WriteFlashcard.h" />
     <ClInclude Include="WriteQuestion.h" />
   </ItemGroup>
